refactor(controller/wavs): extract browser list rebuilding
